Put code for generating payload in a separate module
